Fix how the events are subscribed by using a list of handlers
It's possible to subscribe several times to the same events with
different callbacks. It's handled by the BusClient.events.subscribe
method. The problem with the previous code was that axon is just
handling events with topics (regex). But in Xcraft we needs to hande
each subscribe for a same topic with a different handler. This is the
purpose of this commit.

The unsubscribe method is returned by the subscribe method. The previous
unsubscribe method is now renamed to unsubscribeAll.