Add a way for sending with the token instead of an orcName
It's for special case where the global busClient must be used without a
provided orcName and where the greathall namespace is not appropriate.
The greathall namespace (server side) can not be used in the case of the
hordes because all serveur will have the same namespace. Here we can use
the server token.